requirements.txt
requests==2.32.5
pytest==7.4.0
pytest-mock==3.11.1
python-dotenv==1.0.0
